Switch public API to the RichPosition struct
In order to implement incremental evaluation updates, the flip bits need
to be exposed by the apply_move function. In order to implement it in a
streamlined way, a RichPosition struct is introduced and the public
methods are switched to it.

Using RichPosition may introduce a mild performance regression, so
upgrading from 0.1.1 to 1.0.0 may make sense only in the view of
potential gains from incremental eval or zobrist hash updates.