Implement unchecked apply_move without function indirection
Profiling has shown that >5% of branches are mispredicted, which appears
to cause a 30%-50% performance hit. In order to address that, a new
unstable API is being implemented. At this stage, the apply_move_unchecked
function was added along with a macro which helps avoid the necessity to
pass a function pointer.